+++ src/main/java/lotto/model/Lotto.java
@@ -1,9 +1,15 @@
 package lotto.model; //패키지 변경은 가능
 
+
+import static lotto.util.NumbersValidator.checkDuplicate;
+import static lotto.util.NumbersValidator.checkElementsInRange;
+import static lotto.util.NumbersValidator.validate;
+
 import java.util.Comparator;
 import java.util.List;
 
-public class Lotto { //번호들이 범위 내에 있는지, 중복되지 않는지 검증
+
+public class Lotto { //번호들이 범위 내에 있는지, 중복되지 않는지 검증, 로또 번호를 오름차순으로 정렬
 
     private final List<Integer> numbers; //접근제어자 변경 불가능
 
@@ -14,25 +20,6 @@ public class Lotto { //번호들이 범위 내에 있는지, 중복되지 않는
         checkElementsInRange(numbers);
         numbers.sort(Comparator.naturalOrder());
         this.numbers = numbers;
-    }
-
-    private void validate(List<Integer> numbers) { //기본적으로 6자인 부분만 확인
-        if (numbers.size() != 6) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    // TODO: 추가 기능 구현
-    private void checkDuplicate(List<Integer> numbers) { //중복 체크
-        if (numbers.size() != numbers.stream().distinct().count()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void checkElementsInRange(List<Integer> numbers) { //범위 체크
-        if (numbers.stream().anyMatch(number -> number < 1 || number > 45)) {
-                throw new IllegalArgumentException();
-        }
     }
 
     public List<Integer> getNumbers() {
